Pluralize: drop `PluralCount` newtype, take plain `u64`
Cast `usize`/`u32` etc. at the call site with `as u64`. The 25-line `Into<PluralCount>` wrapper bought callers a missing cast at the cost of an unusual signature; replacing it with a plain `u64` arg matches what every other Rust codebase does and trims the helper to two ~6-line fns. Test-only `chunks` counters in the AI smoke tests flipped from `usize` to `u64` for the same reason (no cast at the format site).

Caught and fixed via `cargo check --tests`: 23 call sites needed `as u64`; clippy + 1942 rust-tests stayed green.

+++ apps/desktop/src-tauri/src/ai/api_keys.rs
@@ -54,7 +54,7 @@ pub fn save(provider_id: &str, api_key: &str) -> Result<(), AiApiKeyError> {
     crate::secrets::store().set(&key, api_key.as_bytes())?;
     info!(
         "AI API key saved for provider {provider_id} ({})",
-        pluralize(key_len, "byte")
+        pluralize(key_len as u64, "byte")
     );
     Ok(())
 }
